[FE] :art: 요약 문자 컴포넌트 생성
