+++ FE/src/components/record/RecordList.jsx
@@ -50,20 +50,8 @@ const AmountP = styled.span`
     width: 160px;
     text-align: center;
 `;
-export function RecordList({list, selected}){
-    const [cookies, setCookie, removeCookie] = useCookies(["year", "month"]);
-    const ll = list.filter(l=>{
-        if(l.timestamp[0] == cookies.year && l.timestamp[1] == cookies.month){
-            if(selected === "All"){
-                return true;
-            } else if(selected === "Revenue"){
-                return l.amount > 0;
-            } else if(selected === "Expenses"){
-                return l.amount < 0;
-            }
-        }
-    })
-    const content = ll.map(l=>{
+export function RecordList({list}){
+    const content = list.map(l=>{
         return <RowWrapper>
             <CategoryP className="ExtraBold18">{l.category}</CategoryP>
             <ItemP className="ExtraBold18">{l.item}</ItemP>
